+++ experimental/agents/ack_builder_agent/__main__.py
@@ -11,13 +11,10 @@
 """Main entry point for the ACK agent CLI."""
 
 import argparse
-import json
 import logging
-import re
 import warnings
 
 from rich.console import Console
-from rich.markdown import Markdown
 from rich.panel import Panel
 from strands import Agent
 from strands.models import BedrockModel
@@ -30,6 +27,7 @@ from ack_builder_agent.tools import (
     verify_build_completion,
 )
 from config.defaults import DEFAULT_MODEL_ID, DEFAULT_REGION, DEFAULT_TEMPERATURE
+from utils.formatting import pretty_markdown
 
 console = Console()
 
@@ -51,14 +49,7 @@ def configure_logging(debug=False):
     warnings.filterwarnings("ignore", category=DeprecationWarning, module="botocore")
 
 
-def pretty_markdown(md) -> Markdown:
-    # If it's a dict, pretty-print as JSON
-    if isinstance(md, dict):
-        md = json.dumps(md, indent=2)
-    elif not isinstance(md, str):
-        md = str(md)
-    clean = re.sub(r"(\n\s*){3,}", "\n\n", (md or "").strip())
-    return Markdown(clean)
+
 
 
 def run_agent_cli():
