+++ experimental/agents/ack_generator_agent/__main__.py
@@ -11,13 +11,10 @@
 """Main entry point for the ACK Generator agent CLI."""
 
 import argparse
-import json
 import logging
-import re
 import warnings
 
 from rich.console import Console
-from rich.markdown import Markdown
 from rich.panel import Panel
 from strands import Agent
 from strands.models import BedrockModel
@@ -38,6 +35,7 @@ from ack_generator_agent.tools import (
     update_service_generator_config,
 )
 from config.defaults import DEFAULT_MODEL_ID, DEFAULT_REGION, DEFAULT_TEMPERATURE
+from utils.formatting import pretty_markdown
 
 console = Console()
 
@@ -57,16 +55,6 @@ def configure_logging(debug=False):
 
     # Suppress deprecation warnings from botocore about datetime.utcnow()
     warnings.filterwarnings("ignore", category=DeprecationWarning, module="botocore")
-
-
-def pretty_markdown(md) -> Markdown:
-    # If it's a dict, pretty-print as JSON
-    if isinstance(md, dict):
-        md = json.dumps(md, indent=2)
-    elif not isinstance(md, str):
-        md = str(md)
-    clean = re.sub(r"(\n\s*){3,}", "\n\n", (md or "").strip())
-    return Markdown(clean)
 
 
 def run_agent_cli():
